+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.25] - 2025-07-19
+
+### Added
+- Added a story for a form example
+
+### Fixed
+- Fixed `ThemeProvider` to correctly use the stored theme in the context
+- Fixed disabled state and styling on `Checkbox`, `Input`, `Select`, `MultiSelect`
+
 ## [0.1.24] - 2025-07-17
 
 ### Fixed